chat2scenario_script
