+++ config/config.json
@@ -1,0 +1,29 @@
+{
+    "name": "Scenario 0",
+    "asam_version": "ASAM OpenSCENARIO V1.2.0",
+    "dataset_option": "highD",
+    "dataset_path_template": "/C:/PhD/Dataset/highD/data/{track_num}_tracks.csv",
+    "metric_option": "Time-Scale",
+    "metric_suboption": "Time To Collision (TTC)",
+    "metric_threshold": "1 - 3",
+    "CA_Input": null,
+    "target_value": null,
+    "openai_key": "*******************************************",
+    "model": "gpt-3.5-turbo",
+    "base_url": "https://api.chatanywhere.org",
+    "output_dir": "./output/",
+    "track_nums": [1],
+    "max_workers": 12,
+    "metric_options": {
+        "Jerk-Scale": {
+            "Longitudinal jerk (LongJ)": "1.5 - 100"
+        },
+        "Time-Scale": {
+            "Time To Collision (TTC)": "0.01 - 2",
+            "Time Headway (THW)": "0.01 - 0.3"
+        },
+        "Acceleration-Scale": {
+            "Deceleration to safety time (DST)": "6 - 100"
+        }
+    }
+}
