+++ config/config_scenario_descriptions.txt
@@ -1,0 +1,3 @@
+1. The ego vehicle is driving straight in the ego lane, and the target vehicle ahead changes lanes to the left adjacent lane.
+2. The ego vehicle is accelerating in the ego lane, and the target vehicle ahead changes lanes to the left adjacent lane.
+# Add more scenario descriptions as needed...
