+++ tutorials/__init__.py
@@ -1,0 +1,1 @@
+# This can be empty, but it's needed to make the tutorials directory a Python package
